Split tests module to factor out test infrastructure
